requirements.txt
flask
firebase-admin
pyrebase4
gunicorn